requirements.txt
chromadb
sentence-transformers
nltk
PyPDF2
pdfplumber
python-docx
google-genai
streamlit